type_traits reworked + clang format
