+++ cpp/oneapi/dal/algo/decision_forest/common.hpp
@@ -16,8 +16,9 @@
 
 #pragma once
 
-#include "oneapi/dal/table/common.hpp"
 #include "oneapi/dal/detail/common.hpp"
+#include "oneapi/dal/table/common.hpp"
+#include "oneapi/dal/util/type_traits.hpp"
 
 namespace oneapi::dal::decision_forest {
