+++ cpp/oneapi/dal/algo/linear_kernel/common.hpp
@@ -16,8 +16,8 @@
 
 #pragma once
 
-#include "oneapi/dal/table/common.hpp"
 #include "oneapi/dal/detail/common.hpp"
+#include "oneapi/dal/table/common.hpp"
 
 namespace oneapi::dal::linear_kernel {
