+++ cpp/oneapi/dal/algo/rbf_kernel/backend/cpu/compute_kernel_test.cpp
@@ -16,8 +16,8 @@
 
 #include "gtest/gtest.h"
 #include "oneapi/dal/algo/rbf_kernel.hpp"
-#include "oneapi/dal/table/row_accessor.hpp"
 #include "oneapi/dal/table/homogen.hpp"
+#include "oneapi/dal/table/row_accessor.hpp"
 
 #include <cmath>
