+++ cpp/oneapi/dal/algo/rbf_kernel/backend/gpu/compute_kernel_dense_dpc_test.cpp
@@ -19,8 +19,8 @@
 #include "gtest/gtest.h"
 #define ONEAPI_DAL_DATA_PARALLEL
 #include "oneapi/dal/algo/rbf_kernel.hpp"
-#include "oneapi/dal/table/row_accessor.hpp"
 #include "oneapi/dal/table/homogen.hpp"
+#include "oneapi/dal/table/row_accessor.hpp"
 
 #include <cmath>
