+++ cpp/oneapi/dal/table/common_type_traits.hpp
@@ -16,25 +16,24 @@
 
 #pragma once
 
+#include "oneapi/dal/table/table_metadata.hpp"
 #include "oneapi/dal/util/detail/type_traits.hpp"
 
 namespace oneapi::dal {
 
 template <typename T>
-inline constexpr bool check_mask_flag(T mask, T flag) {
-    using U = std::underlying_type_t<T>;
-    return (static_cast<U>(mask) & static_cast<U>(flag)) > 0;
-}
+struct is_table_impl {
+    ONEAPI_DAL_SIMPLE_HAS_METHOD_TRAIT(std::int64_t, get_column_count, () const)
+    ONEAPI_DAL_SIMPLE_HAS_METHOD_TRAIT(std::int64_t, get_row_count, () const)
+    ONEAPI_DAL_SIMPLE_HAS_METHOD_TRAIT(const table_metadata&, get_metadata, () const)
+    ONEAPI_DAL_SIMPLE_HAS_METHOD_TRAIT(std::int64_t, get_kind, () const)
+
+    static constexpr bool value = has_method_get_column_count_v<T> &&
+                                  has_method_get_row_count_v<T> && has_method_get_metadata_v<T> &&
+                                  has_method_get_kind_v<T>;
+};
 
 template <typename T>
-inline constexpr T bitwise_and(T lhs_mask, T rhs_mask) {
-    using U = std::underlying_type_t<T>;
-    return static_cast<T>(static_cast<U>(lhs_mask) & static_cast<U>(lhs_mask));
-}
+inline constexpr bool is_table_impl_v = is_table_impl<T>::value;
 
-template <typename T>
-inline constexpr T bitwise_or(T lhs_mask, T rhs_mask) {
-    using U = std::underlying_type_t<T>;
-    return static_cast<T>(static_cast<U>(lhs_mask) | static_cast<U>(lhs_mask));
-}
 } // namespace oneapi::dal
